Created seperate create user tab
